适配ubuntu 22.04

+++ 01-installation-scripts/01-MySQL/01-install-mysql5.7.sh
@@ -29,16 +29,16 @@ src_dir=$(pwd)/00src00
 # 端口
 PORT=3306
 # mysql部署好后，root的默认密码
+# 仅centos有效，ubuntu需要交互式的手动输入root密码
 my_root_passwd=123456
 # mysql版本
 mysql_version=5.7.38
-unit_file_name=mysqld.service
 
 # 部署目录的父目录
 DIR=$(pwd)
 # 部署目录的名字，最终的部署目录为${DIR}/${mysql_dir_name}
 mysql_dir_name=mysql-${mysql_version}
-#****************以上为通用变量*****************************
+
 
 # 脚本执行用户检测
 if [[ $(whoami) != 'root' ]];then
@@ -58,13 +58,12 @@ else
 	exit 99
 fi
 
-
-#****************以下为二进制部署才需要的变量******************
 if [[ $os == 'centos' ]];then
-    # 解压后的名字
-    FILE=mysql-${mysql_version}-linux-glibc2.12-x86_64
     # mysql二进制包名字
-    mysql_tgz=${FILE}.tar.gz
+    mysql_tgz=mysql-${mysql_version}-linux-glibc2.12-x86_64.tar.gz
+    unit_file_name=mysqld.service
+elif [[ $os == 'ubuntu' ]];then
+    unit_file_name=mysql.service
 fi
 #############################################################
 
@@ -74,7 +73,7 @@ function untar_tgz(){
     tar xf $1
     if [ $? -ne 0 ];then
         echo_error 解压出错，请检查！
-        exit 1
+        exit 80
     fi
 }
 
@@ -119,7 +118,7 @@ function download_tar_gz(){
             wget $2
             if [ $? -ne 0 ];then
                 echo_error 下载 $2 失败！
-                exit 1
+                exit 80
             fi
             file_in_the_dir=$(pwd)
             # 返回脚本所在目录，这样这个函数才可以多次使用
@@ -139,7 +138,7 @@ function download_tar_gz(){
                 wget $2
                 if [ $? -ne 0 ];then
                     echo_error 下载 $2 失败！
-                    exit 1
+                    exit 80
                 fi
                 file_in_the_dir=$(pwd)
                 cd ${back_dir}
@@ -180,7 +179,7 @@ function init_account(){
     # mysql启动失败的话退出
     if [ $? -ne 0 ];then
         echo_error mysql启动失败，请查看错误信息
-        exit 2
+        exit 81
     fi
 
     # mysql启动成功后的操作
@@ -209,17 +208,29 @@ function init_account(){
 }
 
 function pre_install(){
-    echo_info 安装依赖
-    yum install -y perl-Data-Dumper perl-JSON libaio libaio-devel
-    # 卸载mariadb
-    mariadb_pkgs=$(rpm -qa | grep -i mariadb)
-    mariadb_pkgs_num=$(echo $mariadb_pkgs | wc -l)
-    if [ $mariadb_pkgs_num -ne 0 ];then
-        echo_info 卸载mariadb相关包中，请耐心等待...
-        for pkg in $mariadb_pkgs
-        do
-            rpm -e --nodeps $pkg
-        done
+    if [[ $os == 'centos' ]];then
+        # 卸载mariadb
+        mariadb_pkgs=$(rpm -qa | grep -i mariadb)
+        mariadb_pkgs_num=$(echo $mariadb_pkgs | wc -l)
+        if [ $mariadb_pkgs_num -ne 0 ];then
+            echo_info 卸载之前的mariadb包，请耐心等待...
+            for pkg in $mariadb_pkgs
+            do
+                rpm -e --nodeps $pkg
+            done
+        fi
+
+        echo_info 安装依赖
+        yum install -y perl-Data-Dumper perl-JSON libaio libaio-devel
+    elif [[ $os == 'ubuntu' ]];then
+        mysql_pkgs_num=$(dpkg -l | grep mysql- | wc -l)
+        if [ $mysql_pkgs_num -ne 0 ];then
+            echo_info 卸载之前的mysql
+            dpkg -l | grep mysql- | awk '{print $2}' | xargs dpkg -P
+        fi
+
+        echo_info 安装相关依赖
+        apt install -y libtinfo5 libmecab2
     fi
 }
 
@@ -230,42 +241,50 @@ function gen_my_cnf() {
     mkdir -p ${DIR}/${mysql_dir_name}/{data,log}
     chown -R mysql:mysql ${DIR}/${mysql_dir_name}
 
-    if [ -f /etc/my.cnf ];then
-        mv /etc/my.cnf /etc/my.cnf_`date +%F`
-        echo_warning 检测到配置文件，已备份为/etc/my.cnf_`date +%F`
+    if [[ $os == 'centos' ]];then
+        my_cnf_file=/etc/my.cnf
+    elif [[ $os == 'ubuntu' ]];then
+        my_cnf_file=/etc/mysql/mysql.conf.d/mysqld.cnf
     fi
 
-    # 生成新的/etc/my.cnf
-    echo_info 初始化/etc/my.cnf
-cat > /etc/my.cnf << EOF
+    if [ -f $my_cnf_file ];then
+        mv $my_cnf_file ${my_cnf_file}_`date +%F`
+        echo_warning 检测到配置文件，已备份为${my_cnf_file}_`date +%F`
+    fi
+
+    # 生成新的配置文件
+    echo_info 初始化${my_cnf_file}
+cat > ${my_cnf_file} << EOF
+# The MySQL  Server configuration file.
 # For advice on how to change settings please see
 # http://dev.mysql.com/doc/refman/5.7/en/server-configuration-defaults.html
 
 [client]
-socket=${DIR}/${mysql_dir_name}/data/mysql.sock
+socket = ${DIR}/${mysql_dir_name}/data/mysql.sock
 
 [mysql]
-default-character-set=utf8mb4
-socket=${DIR}/${mysql_dir_name}/data/mysql.sock
+default-character-set = utf8mb4
+socket = ${DIR}/${mysql_dir_name}/data/mysql.sock
 
 [mysqld]
 #skip_grant_tables
 # 跳过dns解析，提升连接速度
 skip_name_resolve
-port=${PORT}
-socket=${DIR}/${mysql_dir_name}/data/mysql.sock
-basedir=${DIR}/${mysql_dir_name}
-datadir=${DIR}/${mysql_dir_name}/data
-max_connections=200
-character_set_server=utf8mb4
-default_storage_engine=INNODB
-max_allowed_packet=16M
+bind_address = 0.0.0.0
+port = ${PORT}
+socket = ${DIR}/${mysql_dir_name}/data/mysql.sock
+basedir = ${DIR}/${mysql_dir_name}
+datadir = ${DIR}/${mysql_dir_name}/data
+max_connections = 200
+character_set_server = utf8mb4
+default_storage_engine = INNODB
+max_allowed_packet = 16M
 # 不区分大小写
 lower_case_table_names = 1
 # 可以避免一些问题
 sql_mode = 
 
-log_error=${DIR}/${mysql_dir_name}/log/mysqld.log
+log_error = ${DIR}/${mysql_dir_name}/log/mysqld.log
 
 ### binlog日志设置
 binlog_format = ROW
@@ -280,13 +299,23 @@ binlog_cache_size = 4m
 #最大binlog缓存大小
 max_binlog_cache_size = 512m
 #配置serverid
-server_id=1
+server_id = 1
+
+
+# 允许符号链接
+symbolic_links = 1
+
 EOF
+
+    # rpm或deb安装
+    if [[ $software -eq 1 ]];then
+        echo 'pid_file=/var/run/mysqld/mysqld.pid' >> ${my_cnf_file}
+    fi
 }
 
 ########## rpm安装mysql
 function install_by_rpm(){
-    rm -f /var/log/mysqld.log
+    [ -f /var/log/mysqld.log ] && rm -f /var/log/mysqld.log
     download_tar_gz ${src_dir} https://mirrors.aliyun.com/mysql/MySQL-5.7/mysql-${mysql_version}-1.el7.x86_64.rpm-bundle.tar
     cd ${file_in_the_dir}
     untar_tgz mysql-${mysql_version}-1.el7.x86_64.rpm-bundle.tar
@@ -304,11 +333,10 @@ function install_by_rpm(){
         echo_info 已成功安装mysql，即将进行一些优化配置
     else
         echo_error 安装出错，请检查！
-        exit 3
+        exit 82
     fi
 
     gen_my_cnf
-    echo 'pid_file=/var/run/mysqld/mysqld.pid' >> /etc/my.cnf
 
     systemctl start mysqld  # 这里启动是为了生成临时密码
     temp_pass=$(grep 'temporary password' ${DIR}/${mysql_dir_name}/log/mysqld.log | awk '{print $NF}')
@@ -318,20 +346,19 @@ function install_by_rpm(){
 function check_dir() {
     if [ -d $1 ];then
         echo_error 目录 $1 已存在，退出
-        exit 4
+        exit 83
     fi
 }
 
 function install_by_tgz(){
     download_tar_gz ${src_dir} https://mirrors.aliyun.com/mysql/MySQL-5.7/${mysql_tgz}
-    #download_tar_gz ${src_dir} https://mirrors.cloud.tencent.com/mysql/downloads/MySQL-5.7/${mysql_tgz}
     cd ${file_in_the_dir}
     untar_tgz ${mysql_tgz}
 
     pre_install
 
     check_dir ${DIR}/${mysql_dir_name}
-    mv ${FILE} ${DIR}/${mysql_dir_name}
+    mv mysql-${mysql_version}-linux-glibc2.12-x86_64 ${DIR}/${mysql_dir_name}
 
     add_user_and_group mysql
 
@@ -372,11 +399,11 @@ EOF
     echo_info 配置PATH环境变量
     if [ -f /usr/local/bin/mysql ];then
         echo_error /usr/local/bin目录有未删除的mysql相关文件，请检查！
-        exit 5
+        exit 84
     fi
     if [ -f /usr/bin/mysql ];then
         echo_error /usr/bin目录有未删除的mysql相关文件，请检查！
-        exit 6
+        exit 85
     fi
     echo "export PATH=\${PATH}:${DIR}/${mysql_dir_name}/bin" > /etc/profile.d/mysql.sh
     source /etc/profile
@@ -391,14 +418,76 @@ function is_run_mysql() {
     ps -ef | grep "${DIR}/${mysql_dir_name}" | grep -v grep &> /dev/null
     if [ $? -eq 0 ];then
         echo_error 检测到mysql正在运行中，退出
-        exit 7
+        exit 86
     fi
 
     if [ -d ${DIR}/${mysql_dir_name} ];then
         echo_error 检测到目录${DIR}/${mysql_dir_name}，请检查是否重复安装，退出
-        exit 8
+        exit 87
     fi
 }
+
+function install_by_deb() {
+    download_tar_gz ${src_dir} https://mirrors.aliyun.com/mysql/MySQL-5.7/mysql-server_${mysql_version}-1ubuntu18.04_amd64.deb-bundle.tar
+    cd ${file_in_the_dir}
+    ls
+    untar_tgz mysql-server_${mysql_version}-1ubuntu18.04_amd64.deb-bundle.tar
+    # 删除测试套件，不需要
+    rm -f mysql-*test*deb
+
+    pre_install
+
+    echo_info 使用deb包安装mysql
+    dpkg -i mysql-*.deb
+    return_code=$?
+    if [[ $return_code -eq 0 || $return_code -eq 1 ]];then
+        echo_info 已成功安装mysql，即将进行一些优化配置
+    else
+        echo_error 安装出错，请检查！
+        exit 88
+    fi
+
+    # ubuntu安装完mysql会启动，这里先停掉
+    systemctl stop mysql
+
+    gen_my_cnf
+    echo_info 迁移mysql数据目录
+    rm -rf ${DIR}/${mysql_dir_name}/data
+    mv /var/lib/mysql ${DIR}/${mysql_dir_name}/data
+    chown -R mysql:mysql ${DIR}/${mysql_dir_name}/data
+
+    echo_info 禁用apparmor
+    if [ -e /etc/apparmor.d/usr.sbin.mysqld ];then
+        rm -f /etc/apparmor.d/disable/usr.sbin.mysqld
+        ln -s /etc/apparmor.d/usr.sbin.mysqld /etc/apparmor.d/disable/
+        apparmor_parser -R /etc/apparmor.d/usr.sbin.mysqld
+    fi
+
+    echo_info 启动mysql
+    systemctl start ${unit_file_name}
+
+    echo_info 设置所有主机均可访问mysql
+    read -p "请输入刚才设置的root密码：" my_root_passwd
+    mysql -uroot -p"${my_root_passwd}" -e "grant all on *.* to root@'%' identified by '${my_root_passwd}' WITH GRANT OPTION;" &> /dev/null
+    if [ $? -ne 0 ];then
+        echo_error 设置权限失败，请确认输入的密码是否正确
+        exit 96
+    fi
+
+    echo_info 重启mysql
+    systemctl restart ${unit_file_name}
+
+    echo_info mysql已启动成功！相关信息如下：
+    echo -e "\033[37m                  端口：${PORT}\033[0m"
+    echo -e "\033[37m                  账号：root\033[0m"
+    echo -e "\033[37m                  密码：${my_root_passwd}\033[0m"
+
+    echo_info mysql控制命令：
+    echo -e "\033[37m                  启动：systemctl start ${unit_file_name}\033[0m"
+    echo -e "\033[37m                  重启：systemctl restart ${unit_file_name}\033[0m"
+    echo -e "\033[37m                  停止：systemctl stop ${unit_file_name}\033[0m"
+}
+
 
 function install_main_func(){
     read -p "请输入数字选择安装类型（如需退出请输入q）：" software
@@ -423,13 +512,37 @@ function install_main_func(){
                 ;;
         esac
     elif [[ $os == 'ubuntu' ]];then
-        echo ubuntu
+        case $software in
+            1)
+                echo_info 即将使用 deb包 安装mysql
+                # 等待1秒，给用户手动取消的时间
+                sleep 1
+                install_by_deb
+                ;;
+            2)
+                echo_info 即将使用 二进制包 安装mysql
+                sleep 1
+                install_by_tgz_ubuntu
+                ;;
+            q|Q)
+                exit 0
+                ;;
+            *)
+                install_main_func
+                ;;
+        esac
     fi
 }
 
 is_run_mysql
 
-echo -e "\033[31m本脚本支持两种部署方式：\033[0m"
-echo -e "\033[36m[1]\033[32m rpm包部署mysql\033[0m"
-echo -e "\033[36m[2]\033[32m 二进制包部署mysql\033[0m"
+if [[ $os == 'centos' ]];then
+    echo -e "\033[31m本脚本支持两种部署方式：\033[0m"
+    echo -e "\033[36m[1]\033[32m rpm包部署mysql\033[0m"
+    echo -e "\033[36m[2]\033[32m 二进制包部署mysql\033[0m"
+elif [[ $os == 'ubuntu' ]];then
+    echo -e "\033[31m本脚本支持两种部署方式：\033[0m"
+    echo -e "\033[36m[1]\033[32m deb包部署mysql\033[0m"
+    echo -e "\033[36m[2]\033[32m 二进制包部署mysql\033[0m"
+fi
 install_main_func
